Finally, the draw-text! should be workable...
But I still doubt the stability of it, though.
I hope the code could be simplified and easy to
debug. Current draw text logic make me feel ill.
It's not a feeling that I hope to write Lisp.
However, since I'm still quite young with lisp,
I think this should be the best for me right now...