Fixed with draw-text! with line-width settings.
How this was a deadly issue?
Previously, nothing happened because I didn't add label width
restrict to draw label. So it won't trigger line-width restriction.

The reason it happens is current-line-width is caculated by
(+ cursor-u move-u), which when draw text like

  (draw-text! ... 50 50 "text" :line-width 40)

will fail to draw because (+ cursor-u move-u) will always larger
than 40 (of course). therefore, replaced with current-line-width,
and the line-width calculation method is same with text-size.