Add support for `unified-histogram-pane'.
I'm not so satisfied with the code, but I think it is usable right
now. The `unified-histogram-pane' should be same as the
`histogram-pane', the only difference shall be that the
`unified-histogram-pane' is unified with the total counts.

Although I was thinking using `histogram' in `ryo/statistics' to
replace the histograming process in `histogram-pane'. But currently,
the `ryo' package is mostly only for my personal usage. Maybe later I
will consider making it more better.